+++ README.md
@@ -1,2 +1,10 @@
-# capstone
-Repo for capstone project by Team 9 in cohort 5 of the IISc Advanced Data Science course.
+## Automated Answer Evaluation Demo
+
+conda create -n capstone python=3.9 anaconda  
+
+conda activate capstone  
+
+pip install -r requirements.txt  
+
+streamlit run app.py  
+
