+++ data/sciq/readme.txt
