--- demo/README.md
@@ -1,10 +1,0 @@
-## Automated Answer Evaluation Demo
-
-conda create -n capstone python=3.9 anaconda  
-
-conda activate capstone  
-
-pip install -r requirements.txt  
-
-streamlit run app.py  
-
